move moments into algo.__init__
Make test_algo.py and remove tests for removed files
