+++ ncempy/test/test_algo.py
@@ -8,7 +8,24 @@ import numpy as np
 import ncempy.io as nio
 from ncempy.eval.stack_align import stack_align
 from ncempy.algo import stack_align as stack_align_algo
+from ncempy.algo import rebin
+import ncempy.algo as nalgo
+from ncempy.algo import fourier_operations
 
+def test_rebin():
+    aa = np.ones((50, 50), dtype='<u2')
+
+    # Test sum and uint16
+    bb = rebin(aa, 2, funcType='sum')
+
+    assert bb[0, 0] == 4
+
+    aa = np.ones((50, 50), dtype='f')
+
+    # Test mean and float
+    bb = rebin(aa, 2, funcType='mean')
+
+    assert bb[0, 0] == 1.0
 
 @pytest.fixture
 def data_location():
@@ -71,3 +88,41 @@ def test_stack_align_cross(data_location):
     for sh_dynamic, known_shift in zip(shifts0, out_stack_shifts):
         assert sh_dynamic[0] == -known_shift[0]
         assert sh_dynamic[1] == -known_shift[1]
+
+
+def test_moments():
+    """Test on a rotated line"""
+
+    ang = 35  # degree
+    sh = (2, -2)
+
+    # Create a a simple line
+    aa = np.zeros((101, 101))
+    aa[50, 1:-1] = 1
+    aa = fourier_operations.rotateImage(aa, ang * np.pi / 180.)
+    aa = fourier_operations.shiftImage(aa, sh)
+
+    M = nalgo.moments(aa)
+    mc = nalgo.moments_central(aa)
+
+    c = nalgo.centroid(M)  # the center
+    print('centroid = {}'.format(c))
+    for mm, nn in zip(sh, c):
+        assert mm + 50 == round(nn)
+
+    th = nalgo.moment_angle(mc) + np.pi/2.  # in radians
+    assert ang == round(th * 180 / np.pi)
+
+
+def test_moments_dtype():
+    """Test that moments are calculated correctly with uint8 input dtype"""
+    # Create test image
+    YY, XX = np.meshgrid(np.linspace(0, 24, 25), np.linspace(0, 24, 25))
+    RR = np.sqrt((XX - 24)**2 + (YY - 24)**2)
+    RR = 255 * np.abs((RR / RR.max()) - 1)
+
+    c1 = nalgo.centroid(nalgo.moments(RR))
+    c2 = nalgo.centroid(nalgo.moments(RR.astype(np.uint8)))
+
+    assert int(c1[0]) == int(c2[0])
+    assert int(c1[1]) == int(c2[1])
